requirements.txt
flask>=3.0
flask-cors>=4.0
flask-sqlalchemy>=3.1
flask-jwt-extended>=4.6
flask-limiter>=3.5
werkzeug>=3.0
python-dotenv>=1.0

# Testing
pytest>=7.0
pytest-cov>=4.0
